Added forced short covering for overleveraged agents
The agent will now be forced to buy as much as their actually available
cash will allow if they are overleveraged. A complementary procedure for
overleveraged long position has also been introduced, although currently,
there are no conditions under which it may become needed.

Also made sure that inability to trade is not a "graveyard state" for
an agent - trading may be allowed once market conditions are more
favourable