Fixes to the ComplexAgent class
Updated the ComplexAgent class to store positive and negative returns
separately (and use corresponding separate empirical cdfs). This
ensures that the z-value (probability of a move at least as extreme as
the one observed) are close to 1 for large deviations and close to 0 for
"small" deviations. Also updated the StandardComplexAgent class,
adding a uniform multiple to the belief adjustment amount.